feat: toc development, stage 3
